azure_functions/function_app.py
#!/usr/bin/env python3
"""
Azure Function App for L.I.F.E. Platform Marketplace Integration
Handles SaaS fulfillment webhooks from Azure Marketplace

Copyright 2025 - Sergio Paya Borrull
"""

import azure.functions as func
import json
import logging
from typing import Dict, Any

app = func.FunctionApp()

@app.route(route="marketplace-webhook", auth_level=func.AuthLevel.FUNCTION, methods=["POST"])
async def marketplace_webhook(req: func.HttpRequest) -> func.HttpResponse:
    """Handle Azure Marketplace SaaS webhook events"""
    logging.info("🚀 L.I.F.E. Platform marketplace webhook triggered")
    
    try:
        # Parse the webhook payload
        webhook_data = req.get_json()
        logging.info(f"Webhook data received: {json.dumps(webhook_data, indent=2)}")
        
        # Extract subscription details
        subscription_id = webhook_data.get("subscriptionId")
        plan_id = webhook_data.get("planId") 
        action = webhook_data.get("action")
        
        logging.info(f"Processing action: {action} for subscription: {subscription_id}, plan: {plan_id}")
        
        # Process different webhook actions
        if action == "Subscribe":
            result = await provision_life_platform_access(subscription_id, plan_id)
        elif action == "Unsubscribe":
            result = await deprovision_life_platform_access(subscription_id)
        elif action == "ChangePlan":
            result = await update_life_platform_plan(subscription_id, plan_id)
        elif action == "ChangeQuantity":
            result = await update_life_platform_quantity(subscription_id, webhook_data.get("quantity"))
        else:
            logging.warning(f"Unknown webhook action: {action}")
            return func.HttpResponse(
                json.dumps({"status": "error", "message": f"Unknown action: {action}"}),
                status_code=400,
                mimetype="application/json"
            )
        
        logging.info(f"Webhook processing successful: {result}")
        return func.HttpResponse(
            json.dumps({"status": "success", "result": result}),
            status_code=200,
            mimetype="application/json"
        )
        
    except Exception as e:
        logging.error(f"❌ Webhook processing error: {str(e)}")
        return func.HttpResponse(
            json.dumps({"status": "error", "message": str(e)}),
            status_code=500,
            mimetype="application/json"
        )

async def provision_life_platform_access(subscription_id: str, plan_id: str) -> Dict[str, Any]:
    """Provision L.I.F.E. platform access for new customer"""
    logging.info(f"🧠 Provisioning L.I.F.E. access for subscription: {subscription_id}, plan: {plan_id}")
    
    # Map marketplace plans to L.I.F.E. platform capabilities
    plan_capabilities = {
        "life-individual": {
            "neural_processing_quota": 1000,  # EEG processing minutes per month
            "max_concurrent_sessions": 1,
            "advanced_algorithms": False,
            "api_access": "basic"
        },
        "life-professional": {
            "neural_processing_quota": 10000,
            "max_concurrent_sessions": 5,
            "advanced_algorithms": True,
            "api_access": "professional"
        },
        "life-enterprise": {
            "neural_processing_quota": 100000,
            "max_concurrent_sessions": 50,
            "advanced_algorithms": True,
            "api_access": "enterprise"
        }
    }
    
    capabilities = plan_capabilities.get(plan_id, plan_capabilities["life-individual"])
    
    # TODO: Integrate with actual L.I.F.E. platform user management
    # This would connect to your existing L.I.F.E. algorithm core:
    # - Create user account in Azure AD B2C
    # - Provision neural processing capacity in your Venturi Gates system
    # - Enable EEG data processing with your experimentP2L algorithms
    # - Set up personalized learning algorithms
    # - Configure access to campaign_manager and azure_functions_workflow
    
    result = {
        "subscription_id": subscription_id,
        "plan_id": plan_id,
        "life_platform_access": "provisioned",
        "capabilities": capabilities,
        "dashboard_url": f"https://lifecoach-121.com/dashboard?subscription={subscription_id}",
        "api_endpoint": "https://life-functions-app.azurewebsites.net/api/neural-processing",
        "eeg_processing_endpoint": "https://life-functions-app.azurewebsites.net/api/eeg-stream",
        "status": "active",
        "provisioned_at": "2025-11-05T12:00:00Z"
    }
    
    logging.info(f"✅ Successfully provisioned L.I.F.E. platform access: {result}")
    return result

async def deprovision_life_platform_access(subscription_id: str) -> Dict[str, Any]:
    """Deprovision L.I.F.E. platform access for cancelled subscription"""
    logging.info(f"🔄 Deprovisioning L.I.F.E. access for subscription: {subscription_id}")
    
    # TODO: Integrate with L.I.F.E. platform cleanup
    # This would connect to your existing L.I.F.E. systems:
    # - Disable user account gracefully
    # - Archive neural processing data from Venturi Gates
    # - Export learning analytics and EEG analysis results
    # - Revoke API access to experimentP2L algorithms
    # - Clean up campaign_manager data
    # - Send final data export to customer
    
    result = {
        "subscription_id": subscription_id,
        "life_platform_access": "deprovisioned", 
        "data_export_url": f"https://life-functions-app.azurewebsites.net/api/export-data/{subscription_id}",
        "final_report_url": f"https://lifecoach-121.com/reports/final/{subscription_id}",
        "status": "cancelled",
        "deprovisioned_at": "2025-11-05T12:00:00Z"
    }
    
    logging.info(f"✅ Successfully deprovisioned L.I.F.E. platform access: {result}")  
    return result

async def update_life_platform_plan(subscription_id: str, new_plan_id: str) -> Dict[str, Any]:
    """Update L.I.F.E. platform plan for existing customer"""
    logging.info(f"🔄 Updating L.I.F.E. plan for subscription: {subscription_id} to plan: {new_plan_id}")
    
    # Re-provision with new capabilities
    result = await provision_life_platform_access(subscription_id, new_plan_id)
    result["action"] = "plan_updated"
    result["updated_at"] = "2025-11-05T12:00:00Z"
    
    return result

async def update_life_platform_quantity(subscription_id: str, quantity: int) -> Dict[str, Any]:
    """Update L.I.F.E. platform quantity for existing customer"""
    logging.info(f"🔄 Updating L.I.F.E. quantity for subscription: {subscription_id} to quantity: {quantity}")
    
    result = {
        "subscription_id": subscription_id,
        "quantity": quantity,
        "scaling_applied": True,
        "new_concurrent_sessions": quantity * 5,  # Scale concurrent processing
        "updated_at": "2025-11-05T12:00:00Z"
    }
    
    return result

@app.route(route="connection-webhook", auth_level=func.AuthLevel.FUNCTION, methods=["POST"])
async def connection_webhook(req: func.HttpRequest) -> func.HttpResponse:
    """Handle marketplace connection webhook"""
    logging.info("🔗 L.I.F.E. Platform connection webhook triggered")
    
    try:
        connection_data = req.get_json()
        logging.info(f"Connection data: {json.dumps(connection_data, indent=2)}")
        
        # Validate the connection request
        # TODO: Implement marketplace token validation
        # TODO: Verify Azure AD tenant and subscription
        
        response = {
            "status": "connected",
            "platform": "L.I.F.E. Platform",
            "version": "2025.11.05",
            "offer_id": "lifecoach121comlimited1759055603743.life-theory",
            "capabilities": [
                "neuroadaptive_learning",
                "eeg_processing", 
                "real_time_adaptation",
                "personalized_algorithms",
                "venturi_gates_processing",
                "campaign_automation",
                "azure_integration"
            ],
            "endpoints": {
                "dashboard": "https://lifecoach-121.com/dashboard",
                "api": "https://life-functions-app.azurewebsites.net/api",
                "docs": "https://lifecoach-121.com/docs",
                "support": "https://lifecoach-121.com/support"
            },
            "connected_at": "2025-11-05T12:00:00Z"
        }
        
        logging.info(f"✅ Connection established successfully: {response}")
        return func.HttpResponse(
            json.dumps(response),
            status_code=200,
            mimetype="application/json"
        )
        
    except Exception as e:
        logging.error(f"❌ Connection webhook error: {str(e)}")
        return func.HttpResponse(
            json.dumps({"status": "error", "message": str(e)}),
            status_code=500,
            mimetype="application/json"
        )

@app.route(route="health", auth_level=func.AuthLevel.ANONYMOUS, methods=["GET"])
async def health_check(req: func.HttpRequest) -> func.HttpResponse:
    """Health check endpoint for L.I.F.E. Platform marketplace integration"""
    
    health_status = {
        "status": "healthy",
        "platform": "L.I.F.E. Platform Marketplace Integration",
        "version": "2025.11.05",
        "timestamp": "2025-11-05T12:00:00Z",
        "services": {
            "marketplace_webhook": "operational",
            "connection_webhook": "operational", 
            "life_platform_backend": "operational",
            "neural_processing": "operational"
        }
    }
    
    return func.HttpResponse(
        json.dumps(health_status),
        status_code=200,
        mimetype="application/json"
    )