azure_functions/requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
azure-cosmos
aiohttp
requests